updates ts files
