+++ utils/update-static-tests-data.js
@@ -1,15 +1,8 @@
 //
-// Usage:
-//
 // add a new item in response tests:
 //
-//   npm run cli.ts xt fetchTrades BTC/USDT undefined 5 -- -- --response --name "My description"
+//   npm run cli.ts kucoin fetchTrades BTC/USDT undefined 5 -- -- --response --name "My description"
 //
-// update specific symbols, specific currencies or ALL (update all currencies & markets):
-//
-//   node ./utils/static-updater-tests-data.js binance BTC/USDT ETH/USDT
-//                                            binance USDC LTC
-//                                            binance ALL
 
 
 import fs from 'fs';
@@ -70,6 +63,26 @@ function die (errorMessage = undefined, code = 1) {
     console.log (errorMessage || defaultMsg);
     process.exit(code);
 }
+
+async function ccxtClass () {
+    let ccxtRef = undefined;
+    try {
+        // if this script is running from tsx, import untranspiled ccxt
+        ccxtRef = await import ('../ts/ccxt.ts');
+    } catch (e) {
+        ccxtRef = ccxt;
+    }
+    return ccxtRef;
+}
+
+function twoSpacedIndent (jsonStr) {
+    return jsonStr.startsWith('{\n  "');
+}
+
+
+
+
+// #####################################
 
 function add_static_result (requestOrResponse, exchangeId, method, entry, spacesIndent = undefined) {
         
@@ -138,146 +151,6 @@ function prependWhitespace(content, spacesAmountPerIndent, indentAmount) {
 }
   
 
-
-
-const dataContainer = {
-    markets: {
-        path: '',
-        json: {},
-        indent: 4,
-    },
-    currencies: {
-        path: '',
-        json: {},
-        indent: 4,
-    },
-};
-
-function twoSpacedIndent (jsonStr) {
-    return jsonStr.startsWith('{\n  "');
-}
-
-async function update_markets_and_currencies () {
-    try {
-        const exchangeId = args[0];
-        if (!exchangeId) {
-            die ();
-        }
-
-        for (const dataType of ['markets', 'currencies']) {
-            dataContainer[dataType].path = rootDir + `/ts/src/test/static/${dataType}/${exchangeId}.json`;
-            const strMarkets = readFileInit(dataContainer[dataType].path, '{}');
-            dataContainer[dataType].json = JSON.parse(strMarkets);
-            dataContainer[dataType].indent = twoSpacedIndent(strMarkets) ? 2 : 4;
-        }
-        //
-        if (!ccxt.exchanges.includes(exchangeId)) {
-            console.log('Exchange id ' + exchangeId + ' not found in exchanges.json');
-            process.exit(1);
-        }
-    
-        // remove first item
-        args.shift ();
-        const symbolsOrCurrencies = args;
-        if (!symbolsOrCurrencies.length) {
-            die ();
-        }
-
-        const settings = getExchangeSettings (exchangeId);
-        let ccxtRef = undefined;
-        try {
-            // if this script is running from tsx, import untranspiled ccxt
-            ccxtRef = await import ('../ts/ccxt.ts');
-        } catch (e) {
-            ccxtRef = ccxt;
-        }
-        const exchange = new ccxtRef[exchangeId]({ ...settings });
-        let currencies = undefined;
-        if (exchange.has['fetchCurrencies']) {
-            currencies = await exchange.fetchCurrencies();
-        }
-        await exchange.loadMarkets();
-
-        for (const argument of symbolsOrCurrencies) {
-            // if it's update command for markets or currencies
-            if (!argument) {
-                die ();
-            }
-            if (argument === 'ALL') {
-                // reserved keyword to update all markets and currencies
-                updateMarketsOrCurrencies (exchange, 'markets', exchange.markets);
-                updateMarketsOrCurrencies (exchange, 'currencies', currencies);
-            } else {
-                // update signle market or currency
-                updateMarketOrCurrency (exchange, argument);
-            }
-        }
-        // @ts-expect-error
-        die ('Finished! ' + exchangeId + ' > ' + jsonStringify (symbolsOrCurrencies), 0);
-    } catch (e) {
-        // @ts-expect-error
-        die ('Static data writeJson error: ' + e.stack.toString (), 1);
-    }
-}
-
-// update all markets or currencies
-function updateMarketsOrCurrencies (exchange, type, source) {
-    if (!source) { // if undefined, e.g. from fetchCurrencies
-        return;
-    }
-    let targetJson = dataContainer[type].json;
-    // get existing keys which needs update
-    const keys = Object.keys (targetJson);
-    // update all existing keys
-    for (const key of keys) {
-        if (key in source) {
-            targetJson[key] = source[key];
-        } else {
-            // if symbol or currency is removed from exchange
-            console.log ('[info] can not update data for key, it is no longer found in latest fetched ' + exchange.id + ' > ' + type + ' > ' + key);
-        }
-    }
-    writeJson(dataContainer[type].path, targetJson, dataContainer[type].indent);
-}
-
-// update signle market or currency
-function updateMarketOrCurrency (exchange, symbolOrCurrency) {
-    // check whether it's market or currency
-    const isMarketOrCurrency = symbolOrCurrency.includes('/');
-
-    const targetObject = isMarketOrCurrency ? exchange.markets[symbolOrCurrency] : exchange.currencies[symbolOrCurrency];
-    if (!targetObject) {
-        // @ts-expect-error
-        die ('Symbol or Currency ['  + (symbolOrCurrency || 'undefined')  + '] not found in '+ exchange.id, 1);
-    }
-    // writeJson to file
-    if (isMarketOrCurrency) {
-        // if it's market, then writeJson market object and currencies too
-        // market object
-        dataContainer.markets.json[symbolOrCurrency] = targetObject;
-        writeJson (dataContainer.markets.path, dataContainer.markets.json, dataContainer.markets.indent);
-        // base & quote currency objects
-        const base = targetObject.base;
-        const quote = targetObject.quote;
-        const baseCurrency = exchange.currencies[base];
-        const quoteCurrency = exchange.currencies[quote];
-        dataContainer.currencies.json[base] = baseCurrency;
-        dataContainer.currencies.json[quote] = quoteCurrency;
-        writeJson (dataContainer.currencies.path, dataContainer.currencies.json, dataContainer.currencies.indent);
-    } else {
-        // if currency, then only writeJson currency object
-        dataContainer.currencies.json[symbolOrCurrency] = targetObject;
-        writeJson(dataContainer.currencies.path, dataContainer.currencies.json, dataContainer.currencies.indent);
-    }
-}
-
-
 export default {};
 
-export { add_static_result, update_markets_and_currencies };
-
-
-if (process.argv.includes ('--update')) {
-    args.shift ();
-    update_markets_and_currencies();
-}
+export { add_static_result };
